[fixed] fix a bug where binary files were not being properly encoded when proxied
Proxy server is not working properly when it requests url which expected
to take binary data as response. The problem is encoding in request
options is not set properly. According to 'request' package document,
those requests expected for binary data should set encoding to null. Thus
a new parameter added to 'proxy' function is to set proper encoding to
different situation. [default: undefined]