Add a new parameter 'encoding' to 'proxy' function
Proxy server is not wokring properly when it requests url which expected
to take binary data as response. The problem is encoding in request
options is not setted properly. According to 'request' package document,
those requests expected for binary data should set encoding to null. Thus
a new parameter added to 'proxy' function is to set proper encoding to
differnet situation. [default: undefined]